Former-commit-id: e56fc391b48e08769f04b0f8bac03ff401dd73e6

+++ invest_natcap/hydropower/hydropower_water_yield.py
@@ -457,7 +457,7 @@ def execute(args):
     demand_table_file = open(args['demand_table_uri'])
     reader = csv.DictReader(demand_table_file)
     for row in reader:
-        demand_dict[int(row['lucode'])] = int(row['demand'])
+        demand_dict[int(row['lucode'])] = float(row['demand'])
     
     LOGGER.debug('Demand_Dict : %s', demand_dict)
     demand_table_file.close()
